pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lychee-ripe"
version = "0.1.0"
description = "Lychee detection and ripeness API service"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.1",
  "uvicorn[standard]>=0.35.0",
  "pydantic>=2.12.5",
  "python-multipart>=0.0.20",
  "pyyaml>=6.0.2",
  "numpy>=2.4.2",
  "opencv-python-headless>=4.13.0.92",
  "ultralytics>=8.4.14",
]

[dependency-groups]
dev = [
  "pytest>=8.4.1",
  "httpx>=0.28.1",
]

[tool.uv]
default-groups = ["dev"]
package = false

[tool.pytest.ini_options]
pythonpath = ["."]
addopts = "-q"
testpaths = ["tests"]
markers = [
  "perf: performance tests"
]
